# Conflicts:
#	core/src/main/java/com/lantanagroup/link/db/SharedService.java
#	core/src/main/resources/shared-db.sql

+++ api/src/main/java/com/lantanagroup/link/api/controller/QAController.java
@@ -3,19 +3,24 @@ package com.lantanagroup.link.api.controller;
 import com.lantanagroup.link.config.api.ApiConfig;
 import com.lantanagroup.link.db.SharedService;
 import com.lantanagroup.link.db.TenantService;
+import com.lantanagroup.link.db.model.Report;
+import com.lantanagroup.link.model.ReportBase;
 import lombok.Setter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.ApplicationContext;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/qa/{tenantId}")
@@ -54,6 +59,13 @@ public class QAController extends BaseController {
   @DeleteMapping("/$deletePatientListById/{id}")
   public void deletePatientListById(@PathVariable String tenantId, @PathVariable UUID id){
     TenantService tenantService = TenantService.create(this.sharedService, tenantId);
+    List<Report> reports = tenantService.getReportsByPatientListId(id);
+    if (!reports.isEmpty()) {
+      String message = String.format(
+              "The following reports reference the specified patient list: %s",
+              reports.stream().map(ReportBase::getId).collect(Collectors.joining(", ")));
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, message);
+    }
     tenantService.deletePatientListById(id);
   }
 
@@ -64,12 +76,12 @@ public class QAController extends BaseController {
   }
 
   @DeleteMapping("/$deletePatientFromList/{patientListId}/{patientId}")
-  public void deletePatientByIDAndListId(@PathVariable String tenantId, @PathVariable String patientListId, @PathVariable String patientId){
+  public void deletePatientByIDAndListId(@PathVariable String tenantId, @PathVariable UUID patientListId, @PathVariable String patientId){
     TenantService tenantService = TenantService.create(this.sharedService, tenantId);
     tenantService.deletePatientByListAndPatientId(patientId,patientListId);
   }
 
-  @DeleteMapping("/$deleteReport/{reportId}")
+  @DeleteMapping("/$deleteReportById/{reportId}")
   public void deleteReport(@PathVariable String tenantId,@PathVariable String reportId){
     TenantService tenantService = TenantService.create(this.sharedService, tenantId);
     tenantService.deleteReport(reportId);
